feat: category delete added and table updated.
Category delete functionality added to both frontend and backend. Backend category model updated, now it holds references to the products which it has. Repositories updated to reflect the changes for deleting and updating categories. Related endpoints added to backend. If a category gets deleted then all the products are moved to the others category. Others category can not be deleted and edited.