Feature/ffa adminpanelcategoriesscreen 01082022 (#15)
* feat: Admin panel and tabs added, category management added.

Admin panel added with sidebar and tabs. Category add and edit works. Categories can be viewed on table. And some naming refactorings were made.

* feat: category delete added and table updated.

Category delete functionality added to both frontend and backend. Backend category model updated, now it holds references to the products which it has. Repositories updated to reflect the changes for deleting and updating categories. Related endpoints added to backend. If a category gets deleted then all the products are moved to the others category. Others category can not be deleted and edited.

* fix: Fixed the bug that error toast not showing.

Error toast was not popping because error checks on subscriptions made in next function. Subscriptions are updated to use observer objects with next, and error attributes.

* refactor: Removed unused import and comment line.

Removed unused import and comment line

* feat: Active class added to category links.

When a category page visited now it's link gets the active class.