feat: JWT based authentication and authorization added.
A seperate project which is an auth server added. Login, Register and ChangePassword methods are usable. After authentication this server produces a JWT and this JWT can be used to access Shopping API endpoints. Swagger page of Shopping API updated to be able to use JWT. Identities are kept in a seperate database and created another context: IdentityContext. There are 3 roles for identities: Admin, Customer, Seller. Default admin and customer accounts seeded. JWTs are signed with secret JWT key which is hold in server environment. Same key is used for Shopping API to be able to authorize JWT. JWT Claims are: sub, jti, email, roles, exp, iss, aud. Accounts are based on IdentityUser in AspNetCore.