feat: Customer creation in sync with AuthServer added.
Customer create, getAll endpoints added to Shopping API. AuthService service added. AuthService is used to communicate with AuthServer. It uses HttpClient. When necessary information is taken from the customer, a new identity created on AuthServer and this information redirected via httpClient. If identity creation is successfull, new customer is created and identityId added to connect it with the AuthServer identity. If any failure occurs on AuthServer, error responses are redirected to client as they are. Also a custom validation error structure added to be able to return same response structure on every response in case of a model validation error occurs by the data annotation attributes on DTOs.