feat: Event publishing, subscribing mechanism added with RabbitMQ.
RabbitMQ client, Event and EventHandler abstractions and interfaces added. RabbitMQ bus added. To create an event a class should implement the Event class. To handle the event an event handler implementing IEventHandler<T> interface must be created and this event handler must be registered to services as transient. RabbitMQBus is injected to classes and publish/subscribe methods can be used. A common infrastrucre added to project. Event registrations are handled in there. If any service needs to use the bus, it needs to reference to GlobalInfrastructure and call app.AddGlobalInfrastructure(config) method in Program.cs.