fix: General bug fixes and controls.
Every functionality controlled, following bugs fixed. Categories nav items styling fixed and now it goes to a new line if there are many categories that can't fit in one line. Coupon notifications now only comes to logged in customers not to visitors. If product is not active, now product details page navigates to home page. Basket and payment page product images fixed, payment page images does not work for localhost because stripe website can't reach to our machine in localhost obviously, so a default image link was put instead. Category deleting was not working after repository changes now it was fixed. On edit product modal the existing category was not selected as default when the modal opens, now it was fixed.